refactor: switcher と service のディレクトリ整理
